#12 Separate SRN wrapper into ODE and SRN
